refactor: コンポーネントをコロケーション
- ProjectCard → app/projects/
- SeverityChip → app/projects/[id]/
- UploadArea → app/
- TeamSettings → app/settings/
- AppHeader, AppSnackbar は layout で使うため app/components/ に残す
